eduVerse_ai/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
pypdf==4.2.0
python-docx==1.1.2   
openpyxl==3.1.3

python-docx==1.1.2
# DOCX/DOC parsing
docx2txt==0.8         


# Vector DB (local, no API key)
chromadb==0.5.0
sentence-transformers==3.0.1
numpy<2.0.0

# Authentication & Security
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
passlib>=1.7.4